+++ example/spec/parallel/ingestGranule/ingestGranuleQueueSpec.js
@@ -9,7 +9,6 @@ const { URL, resolve } = require('url');
 const {
   Granule,
   Pdr,
-  Provider,
 } = require('@cumulus/api/models');
 const GranuleFilesCache = require('@cumulus/api/lib/GranuleFilesCache');
 const {
@@ -39,6 +38,7 @@ const {
 } = require('@cumulus/integration-tests/api/distribution');
 const { LambdaStep } = require('@cumulus/integration-tests/sfnStep');
 const { getExecution } = require('@cumulus/api-client/executions');
+const { deleteProvider } = require('@cumulus/api-client/providers');
 
 const { waitForApiStatus } = require('../../helpers/apiUtils');
 const {
@@ -92,7 +92,6 @@ describe('The S3 Ingest Granules workflow', () => {
   let pdrModel;
   let postToCmrOutput;
   let provider;
-  let providerModel;
   let publishGranuleExecutionArn;
   let testDataFolder;
   let workflowExecutionArn;
@@ -112,7 +111,6 @@ describe('The S3 Ingest Granules workflow', () => {
       granuleModel = new Granule();
       process.env.system_bucket = config.bucket;
       process.env.ProvidersTable = `${config.stackName}-ProvidersTable`;
-      providerModel = new Provider();
       process.env.PdrsTable = `${config.stackName}-PdrsTable`;
       pdrModel = new Pdr();
 
@@ -247,7 +245,10 @@ describe('The S3 Ingest Granules workflow', () => {
         collectionName: collection.name,
         collectionVersion: collection.version,
       }),
-      providerModel.delete(provider),
+      deleteProvider({
+        prefix: config.stackName,
+        providerId: provider.id,
+      }),
     ]);
   });