+++ packages/api/tests/lambdas/test-sns-consumer.js
@@ -59,7 +59,6 @@ const provider = { id: 'PROV1' };
 let ruleModel;
 
 test.before(async () => {
-  process.env.ProvidersTable = randomString();
   process.env.RulesTable = randomString();
   process.env.stackName = randomString();
   process.env.system_bucket = randomString();