+++ packages/db/src/types/provider.ts
@@ -4,6 +4,7 @@
  * This interface describes a Provider object in postgres compatible format that
  * is ready for write to Cumulus's postgres database instance
  */
+
 export interface PostgresProvider {
   certificate_uri?: string | null,
   cm_key_id?: string | null,